requirements.txt
flask
flask-cors
flask-sqlalchemy
insightface
opencv-python
numpy
pillow
chromadb
python-dotenv
gunicorn